move ui-tests into ui-tests folders
