+++ testcafeRunner.js
@@ -15,7 +15,7 @@ const browserFromArgs = process.argv.slice(2, 3);
 const browserToRun = browserFromArgs.length > 0 ? browserFromArgs : 'chrome:headless';
 
 const testFromArgs = process.argv.slice(3, 4);
-const testsToRun = testFromArgs ? `**/*${testFromArgs}*/*ui.test.js` : '**/*ui.test.js';
+const testsToRun = testFromArgs ? `**/*${testFromArgs}*/ui-tests/*ui.test.js` : '**/*ui.test.js';
 
 let stopOnFirstFail = true;
 let quarantineMode = true;
